+++ examples/username-password/src/routes/+layout.server.ts
@@ -1,3 +1,4 @@
 import { auth } from '$lib/server/lucia';
+import type { ServerLoadEvent } from '@sveltejs/kit';
 
 export const load = auth.handleServerSession()