+++ examples/username-password/src/routes/api/random-number/+server.ts
@@ -1,16 +1,8 @@
-import { auth } from '$lib/server/lucia';
 import type { RequestHandler } from '@sveltejs/kit';
 
-export const GET: RequestHandler = async (event) => {
-	try {
-		await auth.validateRequestEvent(event)
-		const number = Math.floor(Math.random() * 100);
-		return new Response(
-			JSON.stringify({
-				number
-			})
-		);
-	} catch (e) {
+export const GET: RequestHandler = async ({ locals }) => {
+	const session = locals.getSession();
+	if (!session)
 		return new Response(
 			JSON.stringify({
 				error: 'Unauthorized'
@@ -19,5 +11,10 @@ export const GET: RequestHandler = async (event) => {
 				status: 401
 			}
 		);
-	}
+	const number = Math.floor(Math.random() * 100);
+	return new Response(
+		JSON.stringify({
+			number
+		})
+	);
 };