+++ packages/adapter-mongoose/src/t.d.ts
@@ -1,6 +1,7 @@
 /// <reference types="lucia-sveltekit" />
 declare namespace Lucia {
-    interface UserData {}
+    type Auth = any;
+    type UserAttributes = {};
 }
 
 declare namespace App {
@@ -10,6 +11,8 @@ declare namespace App {
 interface UserDoc {
     _id: string;
     __v?: any;
+    _doc?: any;
+    $__?: any;
     hashed_password: string | null;
     provider_id: string;
     username: string;