+++ packages/adapter-mongoose/src/utils.ts
@@ -4,15 +4,17 @@ export const convertUserDoc = (row: UserDoc): UserSchema => {
     const {
         _id: id,
         __v: _,
+        $__,
+        _doc,
         hashed_password,
         provider_id,
-        ...userData
+        ...attributes
     } = row;
     return {
         id,
         hashed_password,
         provider_id,
-        ...userData,
+        ...attributes,
     };
 };