+++ packages/adapter-prisma/src/t.d.ts
@@ -1,8 +1,8 @@
 /// <reference types="lucia-sveltekit" />
 declare namespace Lucia {
-	interface UserData {}
+    type Auth = any;
+    type UserAttributes = {};
 }
-
 
 declare namespace App {
 	interface Locals {}