+++ packages/adapter-supabase/src/index.ts
@@ -94,12 +94,12 @@ const adapter = (url: string, secret: string): Adapter => {
                         id: userId || undefined,
                         provider_id: data.providerId,
                         hashed_password: data.hashedPassword,
-                        ...data.userData,
+                        ...data.attributes,
                     },
                     {
                         returning: "representation",
                     }
-                );
+                ).single();
             if (error) {
                 console.error(error);
                 if (
@@ -113,7 +113,7 @@ const adapter = (url: string, secret: string): Adapter => {
                 }
                 throw new LuciaError("DATABASE_UPDATE_FAILED");
             }
-            return dbData[0].id;
+            return dbData
         },
         deleteUser: async (userId) => {
             const { error } = await supabase