+++ packages/adapter-test/src/ambient.d.ts
@@ -1,8 +1,9 @@
 /// <reference types="lucia-sveltekit" />
 declare namespace Lucia {
-    interface UserData {
-        username: string;
-    }
+    type Auth = any;
+    type UserAttributes = {
+        username: string
+    };
 }
 
 declare namespace App {