+++ packages/adapter-test/src/tests/userid.ts
@@ -14,15 +14,15 @@ export const testAdapter = async (adapter: Adapter, db: Database) => {
         "Insert a user with a null user id into Users DB",
         async () => {
             const user = new User();
-            const userId = await adapter.setUser(null, {
+            const createdUser = await adapter.setUser(null, {
                 providerId: user.providerId,
                 hashedPassword: user.hashedPassword,
-                userData: {
+                attributes: {
                     username: user.username,
                 },
             });
             user.update({
-                id: userId,
+                id: createdUser.id,
             });
             const users = await db.getUsers();
             validate.includesSomeItem(