+++ packages/lucia-sveltekit/src/auth/index.ts
@@ -1,5 +1,5 @@
 import type { Handle } from "../kit.js";
-import type { Env } from "../types.js";
+import type { Env, User } from "../types.js";
 import { handleHooksFunction } from "./hooks.js";
 import {
     authenticateUserFunction,
@@ -7,15 +7,12 @@ import {
     deleteUserFunction,
     getUserFunction,
     getUserByProviderIdFunction,
-    updateUserDataFunction,
+    updateUserAttributesFunction,
     updateUserPasswordFunction,
     updateUserProviderIdFunction,
     getSessionUserFunction,
 } from "./user/index.js";
-import {
-    parseRequestFunction,
-    validateRequestEventFunction,
-} from "./request.js";
+import { parseRequestFunction, validateRequestFunction } from "./request.js";
 import {
     createSessionFunction,
     deleteDeadUserSessionsFunction,
@@ -28,10 +25,10 @@ import {
 import { handleServerSessionFunction } from "./load.js";
 import { deleteAllCookiesFunction } from "./cookie.js";
 import clc from "cli-color";
-import { Adapter } from "../adapter/index.js";
+import { Adapter, UserData, UserSchema } from "../adapter/index.js";
 
-export const lucia = (configs: Configurations) => {
-    return new Auth(configs) as Omit<Auth, "getAuthSession">;
+export const lucia = <C extends Configurations>(configs: C) => {
+    return new Auth(configs) as Omit<Auth<C>, "getAuthSession">;
 };
 
 const validateConfigurations = (configs: Configurations) => {
@@ -46,9 +43,9 @@ const validateConfigurations = (configs: Configurations) => {
     }
 };
 
-export class Auth {
-    private context: Context;
-    constructor(configs: Configurations) {
+export class Auth<C extends Configurations = any> {
+    public context: Context<C>;
+    constructor(configs: C) {
         validateConfigurations(configs);
         this.context = {
             auth: this,
@@ -60,12 +57,27 @@ export class Auth {
             sessionTimeout: configs.sessionTimeout || 1000 * 60 * 60 * 24,
             idlePeriodTimeout:
                 configs.idlePeriodTimeout || 1000 * 60 * 60 * 24 * 14,
+            transformUserData: ({
+                id,
+                hashed_password,
+                provider_id,
+                ...attributes
+            }) => {
+                const transform =
+                    configs.transformUserData ||
+                    (({ id }) => {
+                        return {
+                            userId: id,
+                        };
+                    });
+                return transform({ id, ...attributes }) as User;
+            },
         };
         this.getUser = getUserFunction(this.context);
         this.getUserByProviderId = getUserByProviderIdFunction(this.context);
         this.getSessionUser = getSessionUserFunction(this.context);
         this.createUser = createUserFunction(this.context);
-        this.updateUserData = updateUserDataFunction(this.context);
+        this.updateUserAttributes = updateUserAttributesFunction(this.context);
         this.updateUserProviderId = updateUserProviderIdFunction(this.context);
         this.updateUserPassword = updateUserPasswordFunction(this.context);
         this.deleteUser = deleteUserFunction(this.context);
@@ -84,7 +96,7 @@ export class Auth {
         );
 
         this.parseRequest = parseRequestFunction(this.context);
-        this.validateRequestEvent = validateRequestEventFunction(this.context);
+        this.validateRequest = validateRequestFunction(this.context);
 
         this.handleHooks = handleHooksFunction(this.context);
         this.handleServerSession = handleServerSessionFunction(this.context);
@@ -94,7 +106,9 @@ export class Auth {
     public getUserByProviderId: ReturnType<typeof getUserByProviderIdFunction>;
     public getSessionUser: ReturnType<typeof getSessionUserFunction>;
     public createUser: ReturnType<typeof createUserFunction>;
-    public updateUserData: ReturnType<typeof updateUserDataFunction>;
+    public updateUserAttributes: ReturnType<
+        typeof updateUserAttributesFunction
+    >;
     public updateUserProviderId: ReturnType<
         typeof updateUserProviderIdFunction
     >;
@@ -115,9 +129,7 @@ export class Auth {
     >;
 
     public parseRequest: ReturnType<typeof parseRequestFunction>;
-    public validateRequestEvent: ReturnType<
-        typeof validateRequestEventFunction
-    >;
+    public validateRequest: ReturnType<typeof validateRequestFunction>;
 
     public handleHooks: () => Handle;
     public handleServerSession: ReturnType<typeof handleServerSessionFunction>;
@@ -131,8 +143,20 @@ interface Configurations {
     csrfProtection?: boolean;
     sessionTimeout?: number;
     idlePeriodTimeout?: number;
+    transformUserData?: (userData: UserData) => Record<string, any>;
 }
 
-export type Context = {
-    auth: Auth;
-} & Required<Configurations>;
+export type Context<C extends Configurations = any> = {
+    auth: Auth<C>;
+    adapter: Adapter;
+    env: Env;
+    generateCustomUserId: () => Promise<string | null>;
+    csrfProtection: boolean;
+    sessionTimeout: number;
+    idlePeriodTimeout: number;
+    transformUserData: (
+        userData: UserSchema
+    ) => C["transformUserData"] extends {}
+        ? ReturnType<C["transformUserData"]>
+        : { userId: string };
+};