+++ packages/lucia-sveltekit/src/auth/user/index.ts
@@ -2,4 +2,4 @@ export { authenticateUserFunction } from "./authenticate.js";
 export { createUserFunction } from "./create.js";
 export { deleteUserFunction } from "./delete.js";
 export { getUserFunction, getUserByProviderIdFunction, getSessionUserFunction } from "./get.js";
-export { updateUserProviderIdFunction, updateUserDataFunction, updateUserPasswordFunction } from "./update.js"
+export { updateUserProviderIdFunction, updateUserAttributesFunction, updateUserPasswordFunction } from "./update.js"