+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## v0.13.0
+
+-   [Breaking] `handleHooks()` automatically validates and renew session 
+-   [Breaking] `User` is `{ userId: string }` by default - additional data stored in `user` table are no longer automatically added
+-   [Breaking] Renamed `userData` fields to `attributes` for `createUser()`
+-   [Breaking] Renamed `updateUserData()` to `updateUserAttributes()`
+-   [Breaking] `setUser()` adapter method returns `User` instead of user id
+-   [Breaking] Renamed `Lucia.UserData` to `Lucia.UserAttributes`
+-   [Breaking] Renamed `userData` to `attributes` field for `updateUser()` adapter method
+-   [Breaking] `Lucia.Auth` and `Lucia.UserAttributes` type must be configured for `transformUserData()`
+-   [Breaking] Removed `validateRequestEvent()` method
+-   [Breaking] `setUser()` adapter method returns created `User` instead of user id
+-   Added `validateRequest()` method
+-   New `transformUserData()` config
+
 ## v0.12.0
 
 -   [Breaking] The `session` table should be reset and rebuilt